+++ tests/GlobalPayments.Api.Tests/GpApi/FraudManagementTest.cs
@@ -20,14 +20,10 @@ namespace GlobalPayments.Api.Tests.GpApi
         [ClassInitialize]
         public static void ClassInitialize(TestContext context) {
             ServicesContainer.ConfigureService(new GpApiConfig {
-                AppId = AppIdFraud,
-                AppKey = AppKeyFraud,
+                AppId = AppId,
+                AppKey = AppKey,
                 Environment = Environment.TEST,
                 Channel = Channel.CardNotPresent,
-                AccessTokenInfo = new AccessTokenInfo()
-                {
-                    TransactionProcessingAccountName = "transaction_processing"
-                },
                 RequestLogger = new RequestConsoleLogger(),
                 EnableLogging = true
             });
@@ -76,8 +72,8 @@ namespace GlobalPayments.Api.Tests.GpApi
 
         [TestMethod]
         public void FraudManagementDataSubmissionWithRules() {
-            const string rule1 = "2c49c2e6-5843-4275-9b92-8c9b6dc8e566";
-            const string rule2 = "2cfa3a28-f8f3-42f8-abbf-79b54e35de16";
+            const string rule1 = "0c93a6c9-7649-4822-b5ea-1efa356337fd";
+            const string rule2 = "a539d51a-abc1-4fff-a38e-b34e00ad0cc3";
 
             var rules = new FraudRuleCollection();
             rules.AddRule(rule1, FraudFilterMode.ACTIVE);
@@ -112,11 +108,9 @@ namespace GlobalPayments.Api.Tests.GpApi
         [TestMethod]
         public void FraudManagementDataSubmissionWith_AllRulesActive() {
             var ruleList = new List<string> {
-                "2c49c2e6-5843-4275-9b92-8c9b6dc8e566",
-                "2cfa3a28-f8f3-42f8-abbf-79b54e35de16",
-                "21db158b-4541-4217-aa81-927596465547",
-                "6acbcb2e-79c7-40c3-8c17-b65c5fba2a54",
-                "a7da55fb-69c4-4c41-abb6-c4dded40354e"
+                "0c93a6c9-7649-4822-b5ea-1efa356337fd",
+                "a539d51a-abc1-4fff-a38e-b34e00ad0cc3",
+                "d023a19e-6985-4fda-bb9b-5d4e0dedbb1e"
             };
             
             var rules = new FraudRuleCollection();
@@ -145,11 +139,9 @@ namespace GlobalPayments.Api.Tests.GpApi
         [TestMethod]
         public void FraudManagementDataSubmissionWith_AllRulesOff() {
             var ruleList = new List<string> {
-                "2c49c2e6-5843-4275-9b92-8c9b6dc8e566",
-                "2cfa3a28-f8f3-42f8-abbf-79b54e35de16",
-                "21db158b-4541-4217-aa81-927596465547",
-                "6acbcb2e-79c7-40c3-8c17-b65c5fba2a54",
-                "a7da55fb-69c4-4c41-abb6-c4dded40354e"
+                "0c93a6c9-7649-4822-b5ea-1efa356337fd",
+                "a539d51a-abc1-4fff-a38e-b34e00ad0cc3",
+                "d023a19e-6985-4fda-bb9b-5d4e0dedbb1e"
             };
             
             var rules = new FraudRuleCollection();
@@ -179,11 +171,11 @@ namespace GlobalPayments.Api.Tests.GpApi
 
         [TestMethod]
         public void ReleaseTransactionAfterFraudResultHold() {
+            card.CardHolderName = "Lenny Bruce";
             var trn = card.Charge(98.10m)
                 .WithCurrency(currency)
                 .WithAddress(address)
                 .WithFraudFilter(FraudFilterMode.ACTIVE)
-                .WithCustomerIpAddress("123.123.123.123")
                 .Execute();
 
             Assert.IsNotNull(trn);
@@ -290,11 +282,11 @@ namespace GlobalPayments.Api.Tests.GpApi
 
         [TestMethod]
         public void CaptureTransactionAfterFraudResultHold() {
+            card.CardHolderName = "Lenny Bruce";
             var trn = card.Authorize(10.10m)
                 .WithCurrency(currency)
                 .WithAddress(address)
                 .WithFraudFilter(FraudFilterMode.ACTIVE)
-                .WithCustomerIpAddress("123.123.123.123")
                 .Execute();
 
             Assert.IsNotNull(trn);
@@ -320,11 +312,11 @@ namespace GlobalPayments.Api.Tests.GpApi
         
         [TestMethod]
         public void RefundTransactionAfterFraudResultHold() {
+            card.CardHolderName = "Lenny Bruce";
             var trn = card.Charge(10.10m)
                 .WithCurrency(currency)
                 .WithAddress(address)
                 .WithFraudFilter(FraudFilterMode.ACTIVE)
-                .WithCustomerIpAddress("123.123.123.123")
                 .Execute();
 
             Assert.IsNotNull(trn);
@@ -342,8 +334,8 @@ namespace GlobalPayments.Api.Tests.GpApi
             }
             catch (GatewayException e) {
                 errorFound = true;
-                Assert.AreEqual("Status Code: BadGateway - The refund password you entered was incorrect ", e.Message);
-                Assert.AreEqual("50017", e.ResponseMessage);
+                Assert.AreEqual("Status Code: BadRequest - You can't refund a delayed transaction that has not been sent for settlement You are refunding money to a customer that has not been and never will be charged! ", e.Message);
+                Assert.AreEqual("40087", e.ResponseMessage);
             } finally {
                 Assert.IsTrue(errorFound);
             }
@@ -455,20 +447,14 @@ namespace GlobalPayments.Api.Tests.GpApi
             Assert.AreEqual(TransactionStatus.Captured.ToString().ToUpper(), trn.ResponseMessage);
             Assert.IsNotNull(trn.FraudFilterResponse);
             Assert.AreEqual(FraudFilterResult.RELEASE_SUCCESSFUL.ToString().ToUpper(), trn.FraudFilterResponse.FraudResponseResult);
-            
-            var errorFound = false;
-            try {
-                trn.Refund()
-                    .WithCurrency(currency)
-                    .Execute();
-            }
-            catch (GatewayException e) {
-                errorFound = true;
-                Assert.AreEqual("Status Code: BadGateway - The refund password you entered was incorrect ", e.Message);
-                Assert.AreEqual("50017", e.ResponseMessage);
-            } finally {
-                Assert.IsTrue(errorFound);
-            }
+
+            var refund = trn.Refund()
+                .WithCurrency(currency)
+                .Execute();
+
+            Assert.IsNotNull(refund);
+            Assert.AreEqual("SUCCESS", refund.ResponseCode);
+            Assert.AreEqual(TransactionStatus.Captured.ToString().ToUpper(), refund.ResponseMessage);
         }
         
         [TestMethod]
@@ -871,14 +857,14 @@ namespace GlobalPayments.Api.Tests.GpApi
                     .OrderBy(TransactionSortProperty.TimeCreated)
                     .Where(SearchCriteria.StartDate, startDate)
                     .And(SearchCriteria.EndDate, endDate)
-                    .And(SearchCriteria.RiskAssessmentResult, FraudFilterResult.HOLD)
+                    .And(SearchCriteria.RiskAssessmentResult, FraudFilterResult.PASS)
                     .Execute();
 
             Assert.IsTrue(response.Results.Count > 0);
             
             var trnSummary = response.Results[new Random().Next(1, response.Results.Count)];
             Assert.IsNotNull(trnSummary.FraudManagementResponse);
-            Assert.AreEqual(FraudFilterResult.HOLD.ToString().ToUpper(), trnSummary.FraudManagementResponse.FraudResponseResult);
+            Assert.AreEqual(FraudFilterResult.PASS.ToString().ToUpper(), trnSummary.FraudManagementResponse.FraudResponseResult);
         }
     }
 }